Update EntityPlayer.java.patch
Drones in PneumaticCraft use fake players. After the first initialization it sets the inventory of this fake player to an other. When installed with Battlegear 2 this will cause a crash because of the 'final' marked field. I know Battlegear 2 is doing this to prevent other mods messing with the player's inventory, but could an exception be made for fake players?

Tnx!

MineMaarten